Update first_preprep.sh
This updated script now does the following:

It takes a single argument: the path to the manifest file.
The manifest file should be tab-separated, with R1 paths in the first column and R2 paths in the second column.
It processes each pair of R1 and R2 files listed in the manifest.
It copies the input files to an input directory and generates new reads using iss in an output directory.
After processing all pairs, it concatenates the generated reads and compresses them.

To use this script:

Create a manifest file (e.g., manifest.txt) with tab-separated R1 and R2 paths:
Copy/path/to/sample1_R1.fastq    /path/to/sample1_R2.fastq
/path/to/sample2_R1.fastq    /path/to/sample2_R2.fastq

Save the script to a file (e.g., generate_reads.sh), make it executable with chmod +x generate_reads.sh, and run it like this:
Copy./generate_reads.sh manifest.txt


This script assumes:

The iss tool is installed and available in the system PATH.
The R script is located at ./programs/bin/create_library.R (if not, it will skip this step).
You have write permissions in the current directory to create input and output folders.

Note that this script now uses the input FASTQ files as templates for iss instead of FASTA files. If you need to use FASTA files as input, you'll need to modify the iss generate command accordingly.